Fallback to normal powershell if attempt downgrade fail
Powershell running with "-version 2" to bypass Constrained Language Mode
sometimes fail on newer version of windows due to lack of dependency.
In such cases, the error message is printed to stderr instead and
output from reader is always null

This fix tries to recover from such failure by attempting to run with
default (higher) version of powershell. If that still fails then the
input is the same as of previous version where all elements of the array
is null